requirements.txt
crewai
streamlit
openai
PyPDF2
crewai_tools
databricks-sdk
python-docx
python-docx
python-Levenshtein
fuzzywuzzy
scikit-learn
fpdf
python-dotenv

